% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TargetByLasso-package.R
\name{TargetPre}
\alias{TargetPre}
\title{TargetByLasso}
\usage{
TargetPre(
  single.effect,
  combo.effect,
  ref.matrix,
  single.select.times = 100,
  single.nfold = 10,
  combo.select.times = 10,
  combo.nfold = 10,
  model = 1,
  n.cores = detectCores()
)
}
\arguments{
\item{single.effect}{Single drugs with effects that can be binomial or continue values}

\item{combo.effect}{Combination drugs with effects, only allowing binomial here}

\item{ref.matrix}{A matrix from STITCH, containing all drug-target information}

\item{single.select.times}{Repeat times for single feature selection, the default value is 100}

\item{single.nfold}{Fold number for single selection cross-validation, the default value is 10}

\item{combo.select.times}{Repeat times for combination features selection, the default value is 10}

\item{combo.nfold}{Fold number for combination features selection, the default value is 10}

\item{model}{The function has two models for combo selection 1 = glmnet, 2 = biglasso, the default is glmnet}

\item{n.cores}{Number of cores of CPU asking for tasks running}
}
\description{
This package is used for key target prediction of a disease by employing the results of drug test matrix
}
\keyword{diseases,}
\keyword{key}
\keyword{prediction}
\keyword{prediction,}
\keyword{relationship}
\keyword{systemtic}
\keyword{target}
